[기능개선][3,4,5단계] Money 자료형 long 수정

+++ src/main/java/domain/Money.java
@@ -2,14 +2,14 @@ package domain;
 
 public class Money {
 
-  private final int purchaseMoney;
+  private final long purchaseMoney;
 
-  public Money(int purchaseMoney) {
+  public Money(long purchaseMoney) {
     validatePurchaseMoney(purchaseMoney);
     this.purchaseMoney = purchaseMoney;
   }
 
-  private static void validatePurchaseMoney(int purchaseMoney) {
+  private static void validatePurchaseMoney(long purchaseMoney) {
     if (purchaseMoney < Lottos.MINIMUM_PURCHASE_MONEY) {
       throw new IllegalArgumentException(
           "구입 금액은 " + Lottos.MINIMUM_PURCHASE_MONEY + "원 이상이어야 합니다."
@@ -17,11 +17,11 @@ public class Money {
     }
   }
 
-  public int getPurchaseMoney() {
+  public long getPurchaseMoney() {
     return purchaseMoney;
   }
 
-  public int getTicketCount(int ticketPrice) {
-    return purchaseMoney / ticketPrice;
+  public int getTicketCount(long ticketPrice) {
+    return (int)(purchaseMoney / ticketPrice);
   }
 }
